refactor: move AuditFinding vocabulary into homeboy-audit-contract (#8425)
The biggest audit type, and the one wrongly written off as by-design coupling.
AuditFinding is a pure, dependency-free enum (170 lines: every finding kind, plus
its snake_case serde / FromStr / all_names helpers) with ~81 consumers across the
codebase — the audit engine that produces findings, the refactor engine that
fixes them, the CLI that renders them. It lived inside code_audit/conventions but
is a shared vocabulary, not audit-internal.

Move the enum + its two inherent impls into homeboy-audit-contract as
finding_kind and re-export AuditFinding at the crate root. conventions +
code_audit re-export it so all ~81 consumers keep resolving unchanged; refactor's
direct imports point at the contract.

One wrinkle: findings.rs had a third `impl AuditFinding` (`confidence()` ->
FindingConfidence, a core-side enum). The orphan rule forbids an inherent impl on
a now-foreign type, so confidence() becomes the free function
finding_confidence(&AuditFinding); its 3 call sites (findings, report,
refactor::auto::policy) updated.

This substantially cuts the refactor <-> code_audit coupling (AuditFinding was 14
of refactor's edges). Contrary to the earlier "cycle by design" read, most of
refactor->audit is contract-able vocabulary + output types (CodeAuditResult,
AuditSummary next), reopening the path to decoupling audit.

Verified: contract + core compile clean; 10 findings tests pass; cargo check
--workspace --tests green.

Refs #8425
